+++ src/main/java/in/oneton/idea/spring/assistant/plugin/misc/GenericUtil.java
@@ -1,16 +1,16 @@
 package in.oneton.idea.spring.assistant.plugin.misc;
 
+import com.intellij.application.options.CodeStyle;
 import com.intellij.codeInsight.completion.InsertionContext;
 import com.intellij.lang.jvm.types.JvmPrimitiveTypeKind;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.codeStyle.CodeStyleSettings;
-import com.intellij.psi.codeStyle.CodeStyleSettingsManager;
 import com.intellij.psi.codeStyle.CommonCodeStyleSettings;
 import in.oneton.idea.spring.assistant.plugin.suggestion.SuggestionNode;
 import lombok.experimental.UtilityClass;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.yaml.YAMLLanguage;
 import org.jetbrains.yaml.psi.YAMLKeyValue;
 
 import java.text.BreakIterator;
@@ -122,10 +122,10 @@ public class GenericUtil {
 
   @NotNull
   public static String getCodeStyleIntent(InsertionContext insertionContext) {
-    final CodeStyleSettings currentSettings =
-        CodeStyleSettingsManager.getSettings(insertionContext.getProject());
-    final CommonCodeStyleSettings.IndentOptions indentOptions =
-        currentSettings.getIndentOptions(insertionContext.getFile().getFileType());
+    final @NotNull CommonCodeStyleSettings currentSettings =
+        CodeStyle.getLanguageSettings(insertionContext.getFile(), YAMLLanguage.INSTANCE);
+    final CommonCodeStyleSettings.IndentOptions indentOptions = currentSettings.getIndentOptions();
+    assert indentOptions != null;
     return indentOptions.USE_TAB_CHARACTER ?
         "\t" :
         StringUtil.repeatSymbol(' ', indentOptions.INDENT_SIZE);